api/requirements.txt
numpy
google-api-python-client
google-cloud-storage
pandas
tables